# Added performance improvements