--- a/lib/rc4.rb
+++ b/lib/rc4.rb
@@ -2,40 +2,46 @@ class RC4
   def initialize(str)
     begin
       raise SyntaxError, "RC4: Key supplied is blank"  if str.eql?('')
-
+      initialize_state(str)
       @q1, @q2 = 0, 0
-      @key = []
-      str.each_byte { |elem| @key << elem } while @key.size < 256
-      @key.slice!(256..@key.size-1) if @key.size >= 256
-      @s = (0..255).to_a
-      j = 0
-      0.upto(255) do |i|
-        j = (j + @s[i] + @key[i] ) % 256
-        @s[i], @s[j] = @s[j], @s[i]
-      end
     end
   end
 
   def encrypt!(text)
-    process text
+    index = 0
+    while index < text.length
+      @q1 = (@q1 + 1) % 256
+      @q2 = (@q2 + @state[@q1]) % 256
+      @state[@q1], @state[@q2] = @state[@q2], @state[@q1]
+      text.setbyte(index, text.getbyte(index) ^ @state[(@state[@q1] + @state[@q2]) % 256])
+      index += 1
+    end
+    text
   end
 
+  alias_method :decrypt!, :encrypt!
+
   def encrypt(text)
-    process text.dup
+    encrypt!(text.dup)
   end
 
   alias_method :decrypt, :encrypt
 
   private
 
-  def process(text)
-    text.unpack("C*").map { |c| c ^ round }.pack("C*")
+  # The initial state which is then modified by the key-scheduling algorithm
+  INITIAL_STATE = (0..255).to_a
+
+  # Performs the key-scheduling algorithm to initialize the state.
+  def initialize_state(key)
+    i = j = 0
+    @state = INITIAL_STATE.dup
+    key_length = key.length
+    while i < 256
+      j = (j + @state[i] + key.getbyte(i % key_length)) % 256
+      @state[i], @state[j] = @state[j], @state[i]
+      i += 1
+    end
   end
 
-  def round
-    @q1 = (@q1 + 1) % 256
-    @q2 = (@q2 + @s[@q1]) % 256
-    @s[@q1], @s[@q2] = @s[@q2], @s[@q1]
-    @s[(@s[@q1]+@s[@q2]) % 256]
-  end
 end