### Implementation of the proposed changes of issue #12

State initialization
* Extracted it into its own method
* Avoid unnecessary Range to Array conversion
* Using String#getbyte for faster access to bytes of the key string

Algorithm itself
* Not using String#unpack and Array#pack for accessing data bytes
  but using faster String#getbyte and String#setbyte
* Avoids invocation of a block and a method for each byte by
  inlining the code

Simple benchmark:

~~~ ruby
require 'benchmark/ips'
require 'rc4old'
require 'rc4'

Benchmark.ips do |results|
  text = (0..5000).map { rand(256).chr}.join
  key = 'mykey'
  results.report('old RC4') { RC4Old.new(key).encrypt(text) }
  results.report('new RC4') { RC4.new(key).encrypt(text) }
  results.compare!
end
~~~

And its results:

~~~
Calculating -------------------------------------
             old RC4        36 i/100ms
             new RC4        42 i/100ms
-------------------------------------------------
             old RC4      362.8 (±0.8%) i/s -       1836 in   5.060957s
             new RC4      423.6 (±0.5%) i/s -       2142 in   5.056711s

Comparison:
             new RC4:      423.6 i/s
             old RC4:      362.8 i/s - 1.17x slower
~~~