@caiges 2015-04-01
Your changes were much simpler than I originally thought so I went ahead and merged. Cheers mate!